requirements.txt
opencv-python
torch>=1.8.0
torchvision>=0.9.0
timm>=0.6.0
numpy
Pillow
matplotlib
requests
